Finalizando Atividades do Room,Database criado

+++ app/src/main/java/alura/com/gringotts/data/repositories/home/AccountStatementRepository.kt
@@ -1,21 +1,22 @@
 import alura.com.gringotts.data.AccountStatementDatabase
 import alura.com.gringotts.data.api.ApiInterface
 import alura.com.gringotts.data.models.home.Transaction
-import alura.com.gringotts.data.networkBoundResource
 import alura.com.gringotts.data.session.SessionManager
-import androidx.room.withTransaction
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
 
 class AccountStatementRepository(
     private val sessionManager: SessionManager,
     private val apiInterface: ApiInterface,
-    private val db: AccountStatementDatabase
+    db: AccountStatementDatabase
 ) {
 
     private val transactionDAO = db.accountStatementDAO()
 
-    suspend fun getAccountStatement(initialDate: String, finalDate: String): List<Transaction> {
+    suspend fun getAccountStatement(
+        initialDate: String,
+        finalDate: String
+    ): List<Transaction> {
         return withContext(Dispatchers.IO) {
             val token = sessionManager.getTokens()!!.tokenAuthentication
             val response = apiInterface.transactions(
@@ -23,22 +24,19 @@ class AccountStatementRepository(
                 finalDate,
                 token
             )
-            networkBoundResource(
-                query = {
-                    transactionDAO.getAllTransactions()
-                },
-                fetch = {
-                    response.body()
-                },
-                saveFetchResult = { transaction ->
-                    db.withTransaction {
-                        transactionDAO.deleteAllTransactions()
-                        if (transaction != null) {
-                            transactionDAO.insertTransactions(transaction)
-                        }
-                    }
-                }
-            )
+
+            if(response.code() == 2000) { //Se n tiver mudado nada e a api falar isso
+                transactionDAO.getAllTransactions()
+            }
+            if(response.code() == 1000){ //Se n tiver dado
+                transactionDAO.insertTransactions(response.body()!!)
+            }
+            if(response.code() == 5000) { //Se a api tiver retornado que a resposta do servidor mudou e precisa atualizar o db
+                transactionDAO.deleteAllTransactions()
+                transactionDAO.insertTransactions(response.body()!!)
+
+            }
+
             if (response.isSuccessful) {
                 return@withContext response.body()!!
             } else {
